requirements.txt
Pyrogram==2.0.27
pytube
telegraph
heroku3
aiohttp
aiofiles
hachoir
Pillow
bs4
covid
pySmartDL
SQLAlchemy==1.3.23
TgCrypto
httpx
pytz
psycopg2
CurrencyConverter
deep-translator
gTTS
git-python
Wikipedia-API
speedtest-cli==2.1.3
qrcode
bing-image-downloader
pyfy
uvloop
cinemagoer
moviepy
